move npu_vm_runtime_api from compiler_adapter to utils
