+++ src/plugins/intel_npu/src/compiler_adapter/include/dynamic_graph.hpp
@@ -9,7 +9,7 @@
 #include "intel_npu/common/idynamic_graph.hpp"
 #include "intel_npu/network_metadata.hpp"
 #include "intel_npu/utils/zero/zero_init.hpp"
-#include "npu_vm_runtime_api.hpp"
+#include "intel_npu/utils/vm/npu_vm_runtime_api.hpp"
 #include "openvino/runtime/so_ptr.hpp"
 
 namespace intel_npu {
